+++ Source/Core/SectionModel.swift
@@ -32,6 +32,12 @@ public class SectionModel : Section, SupplementaryAccessible
     /// - Warning: If you try to set new array to this property [T], the only way to do this without exception is to wrap it into items.map { $0 }. This is a workaround that exists because of Swift inability to cast [T] to [Any]. You can call `setItems` method instead of doing so.
     /// - SeeAlso: `setItems:`
     public var items = [Any]()
+    
+    public weak var sectionLocationDelegate: SectionLocationIdentifyable?
+    
+    public var currentSectionIndex: Int? {
+        return sectionLocationDelegate?.sectionIndex(for: self)
+    }
 
     /// Supplementaries dictionary.
     public var supplementaries = [String:[IndexPath:Any]]()